--- calc_100.sh
@@ -1,4 +1,0 @@
-#!/usr/bin/env bash
-
-LANGS
-for
