--- experiments/poleval19/poleval-small.csv
@@ -1,21 +1,0 @@
-,model_dir_parent,model_name,test Accuracy,test F1 score bin,test Loss,test Precision,test Recall,valid Accuracy,valid F1 score bin,valid Loss,valid Precision,valid Recall
-0,data/hate/pl-10-wiki/models/sp25k,lstm_small_ft6_el20_lmseed-0-ftseed-0-clsweightseed-0-clstrainseed-0.m,0.8830000162124634,0.4507042169570923,0.5523807,0.607594907283783,0.35820895433425903,0.9263681769371033,0.593406617641449,0.3514958,0.5567010045051575,0.6352941393852234
-1,data/hate/pl-10-wiki/models/sp25k,lstm_small_ft6_el20_lmseed-0-ftseed-0-clsweightseed-1-clstrainseed-0.m,0.8820000290870667,0.4913793206214905,0.51190376,0.581632673740387,0.4253731369972229,0.9154228568077087,0.5454545617103577,0.3744005,0.5,0.6000000238418579
-2,data/hate/pl-10-wiki/models/sp25k,lstm_small_ft6_el20_lmseed-0-ftseed-0-clsweightseed-2-clstrainseed-0.m,0.8809999823570251,0.4757709205150604,0.5393214,0.5806451439857483,0.4029850661754608,0.9174129366874695,0.5561497807502747,0.37151873,0.5098039507865906,0.6117647290229797
-3,data/hate/pl-10-wiki/models/sp25k,lstm_small_ft6_el20_lmseed-0-ftseed-0-clsweightseed-3-clstrainseed-0.m,0.8740000128746033,0.39423078298568726,0.58161116,0.5540540814399719,0.305970162153244,0.9243780970573425,0.5632184147834778,0.3464281,0.550561785697937,0.5764706134796143
-4,data/hate/pl-10-wiki/models/sp25k,lstm_small_ft6_el20_lmseed-0-ftseed-0-clsweightseed-4-clstrainseed-0.m,0.8799999952316284,0.5238094925880432,0.4988901,0.5593220591545105,0.49253731966018677,0.9164178967475891,0.5670102834701538,0.3931842,0.5045871734619141,0.6470588445663452
-5,data/hate/pl-10-wiki/models/sp25k,lstm_small_ft6_el20_lmseed-0-ftseed-0-clsweightseed-5-clstrainseed-0.m,0.8809999823570251,0.44131454825401306,0.5540554,0.594936728477478,0.3507462739944458,0.9223880767822266,0.5568182468414307,0.36945158,0.5384615659713745,0.5764706134796143
-6,data/hate/pl-10-wiki/models/sp25k,lstm_small_ft6_el20_lmseed-0-ftseed-0-clsweightseed-6-clstrainseed-0.m,0.878000020980835,0.48739495873451233,0.5654785,0.557692289352417,0.43283581733703613,0.9213930368423462,0.5729729533195496,0.38207418,0.5299999713897705,0.6235294342041016
-7,data/hate/pl-10-wiki/models/sp25k,lstm_small_ft6_el20_lmseed-0-ftseed-0-clsweightseed-7-clstrainseed-0.m,0.8809999823570251,0.4803493320941925,0.5358918,0.5789473652839661,0.41044774651527405,0.9174129366874695,0.5414364337921143,0.38049248,0.5104166865348816,0.5764706134796143
-8,data/hate/pl-10-wiki/models/sp25k,lstm_small_ft6_el20_lmseed-0-ftseed-0-clsweightseed-8-clstrainseed-0.m,0.8939999938011169,0.5137615203857422,0.56349623,0.6666666865348816,0.41791045665740967,0.9243780970573425,0.5632184147834778,0.34343374,0.550561785697937,0.5764706134796143
-9,data/hate/pl-10-wiki/models/sp25k,lstm_small_ft6_el20_lmseed-0-ftseed-0-clsweightseed-9-clstrainseed-0.m,0.8830000162124634,0.440191388130188,0.54497576,0.6133333444595337,0.34328359365463257,0.9253731369972229,0.5562130212783813,0.37040198,0.5595238208770752,0.5529412031173706
-10,data/hate/pl-10-wiki/models/sp25k,lstm_small_ft6_el20_lmseed-1-ftseed-0-clsweightseed-0-clstrainseed-0.m,0.8939999938011169,0.526785671710968,0.5066931,0.6555555462837219,0.44029849767684937,0.9243780970573425,0.5681818723678589,0.3419854,0.5494505763053894,0.5882353186607361
-11,data/hate/pl-10-wiki/models/sp25k,lstm_small_ft6_el20_lmseed-1-ftseed-0-clsweightseed-1-clstrainseed-0.m,0.8849999904632568,0.5064377784729004,0.48513892,0.5959596037864685,0.44029849767684937,0.9203979969024658,0.550561785697937,0.39884716,0.5268816947937012,0.5764706134796143
-12,data/hate/pl-10-wiki/models/sp25k,lstm_small_ft6_el20_lmseed-1-ftseed-0-clsweightseed-2-clstrainseed-0.m,0.8730000257492065,0.5605536103248596,0.43568474,0.522580623626709,0.60447758436203,0.9074627161026001,0.5714285969734192,0.3917941,0.469696968793869,0.729411780834198
-13,data/hate/pl-10-wiki/models/sp25k,lstm_small_ft6_el20_lmseed-1-ftseed-0-clsweightseed-3-clstrainseed-0.m,0.8769999742507935,0.5323194265365601,0.5091572,0.5426356792449951,0.5223880410194397,0.9134328365325928,0.5445026159286499,0.40174288,0.49056604504585266,0.6117647290229797
-14,data/hate/pl-10-wiki/models/sp25k,lstm_small_ft6_el20_lmseed-1-ftseed-0-clsweightseed-4-clstrainseed-0.m,0.8799999952316284,0.4957982897758484,0.47033572,0.567307710647583,0.44029849767684937,0.9144278764724731,0.5520833730697632,0.36492744,0.4953271150588989,0.6235294342041016
-15,data/hate/pl-10-wiki/models/sp25k,lstm_small_ft6_el20_lmseed-1-ftseed-0-clsweightseed-5-clstrainseed-0.m,0.8849999904632568,0.5267489552497864,0.46872956,0.5871559381484985,0.4776119291782379,0.9094527363777161,0.5380710959434509,0.38632104,0.4732142984867096,0.6235294342041016
-16,data/hate/pl-10-wiki/models/sp25k,lstm_small_ft6_el20_lmseed-1-ftseed-0-clsweightseed-6-clstrainseed-0.m,0.8859999775886536,0.4999999403953552,0.49172023,0.6063829660415649,0.4253731369972229,0.9194029569625854,0.5423728227615356,0.36051777,0.52173912525177,0.5647059082984924
-17,data/hate/pl-10-wiki/models/sp25k,lstm_small_ft6_el20_lmseed-1-ftseed-0-clsweightseed-7-clstrainseed-0.m,0.878000020980835,0.5040649771690369,0.49449727,0.5535714030265808,0.46268656849861145,0.9134328365325928,0.5445026159286499,0.41053435,0.49056604504585266,0.6117647290229797
-18,data/hate/pl-10-wiki/models/sp25k,lstm_small_ft6_el20_lmseed-1-ftseed-0-clsweightseed-8-clstrainseed-0.m,0.8859999775886536,0.5648854970932007,0.41943383,0.578125,0.5522388219833374,0.9074627161026001,0.5230768918991089,0.34020123,0.4636363685131073,0.6000000238418579
-19,data/hate/pl-10-wiki/models/sp25k,lstm_small_ft6_el20_lmseed-1-ftseed-0-clsweightseed-9-clstrainseed-0.m,0.8899999856948853,0.5416666865348816,0.481508,0.6132075190544128,0.48507463932037354,0.9094527363777161,0.5081080794334412,0.4094172,0.4699999988079071,0.5529412031173706
