--- ulmfit/pretrain_lm.py
@@ -1,413 +1,0 @@
-"""
-Script to train a model on a preprocessed Wiki dataset. Note that the dataset is
-expected to have been tokenized with Moses and processed with `postprocess_wikitext.py`.
-That is, the data is expected to be white-space separated and numbers are expected
-to be split.
-"""
-import pathlib
-from dataclasses import asdict
-from string import Template
-
-import fastai
-
-from fastai import *
-from fastai.callbacks import CSVLogger
-import fastai.text
-from fastai.text import *
-import torch
-from ulmfit.datasets.utils import read_whitespace_file, \
-    validate, UNK
-from fastai_contrib.text_data import MosesPreprocessingFunc, get_sentencepiece, get_sentencepiece_fastai, \
-    make_data_bunch_from_df
-import pickle
-
-from pathlib import Path
-
-LM_BEST = "lm_best"
-ENC_BEST = "enc_best"
-
-
-class Tokenizers(Enum):
-    FASTAI_SUBWORD = 'fsp'
-    SUBWORD='sp'
-    MOSES='v'
-    MOSES_FA='vf'
-    FASTAI='f'
-
-def istitle(line):
-    return len(re.findall(r'^ ?= [^=]* = ?$', line)) != 0
-
-def read_wiki_articles(filename):
-    if "reddit" in str(filename): # Temporary hack to handle poleval reddit dataset
-        return pd.read_csv(filename, header=None, names=["texts"]).fillna("")
-
-    articles = []
-
-    with open(filename, encoding='utf8') as f:
-        lines = f.readlines()
-    current_article = []
-    for i,line in enumerate(lines):
-        current_article.append(line)
-        if i < len(lines)-2 and lines[i+1].strip() == "" and istitle(lines[i+2]):
-            articles.append("".join(current_article))
-            current_article = []
-    articles.append("".join(current_article))
-    print(f"Wiki text was split to {len(articles)} articles")
-    return pd.DataFrame({'texts': np.array(articles, dtype=np.object)})
-
-
-name_re = re.compile("(bwd)?(lstm|qrnn)_(.*)_(lmseed-)?.*\.m")
-def folder_name_to_model_name(folder_name):
-    if hasattr(folder_name, 'name'):
-        folder_name = folder_name.name
-    match = name_re.match(folder_name)
-    if match:
-        return match.group(3)
-    return None
-
-
-@dataclass
-class DataSetParams:
-    dataset_path: str  # data_dir
-    lang: str = None
-
-    def __post_init__(self):
-        if self.lang is None:
-            self.lang = infer_lang_from_dataset(self.dataset_path.name)
-
-    def resolve_template(self, template, **additional_options):
-        try:
-            params = asdict(self)
-            params.update(additional_options)
-            params["dataset_name"] = self.dataset_path.name
-            return Template(template).substitute(**params)
-        except KeyError as e:
-            raise KeyError(f"{e} , options:{repr(list(params.keys()))}")
-
-@dataclass
-class LMHyperParams(DataSetParams):
-    base_lm_path: str = None
-    backwards: str = False
-    bidir: bool = False
-    qrnn: bool = True
-    max_vocab: int = 60000
-    tokenizer: Tokenizers = Tokenizers.MOSES
-    pretrained_model: str = None
-
-    emb_sz:int = 400
-    nh: int = None
-    nl: int = 3
-    out_bias: bool = True
-
-    lmseed: int = None
-
-    clip: float = 0.12
-    bptt: int = 70
-    # alpha and beta - defaults like in fastai/text/learner.py:RNNLearner()
-    rnn_alpha: float = 2  # activation regularization (AR)
-    rnn_beta: float = 1  # temporal activation regularization (TAR)
-
-    name: str = None
-    cuda_id: InitVar[int] = 0
-    tokenizer_mod: str = ''
-
-    def __post_init__(self, cuda_id):
-        if self.bidir and self.backwards:
-            raise ValueError('Both "backwards" and "bidir" options cannot be enabled at the same time')
-        if not torch.cuda.is_available():
-            print('CUDA not available. Setting device=-1.')
-            cuda_id = -1
-        torch.cuda.set_device(cuda_id)
-        self.dataset_path = Path(self.dataset_path)
-        self.base_lm_path = Path(self.base_lm_path) if self.base_lm_path is not None else None
-        self.tokenizer = Tokenizers(self.tokenizer) if isinstance(self.tokenizer, str) else self.tokenizer
-
-        assert self.dataset_path.exists(), f"The dataset_path {self.dataset_path} does not exists"
-        self.cache_dir = self.dataset_path / 'models' / self.tokenizer_prefix
-        self.model_path = self.cache_dir / self.model_name
-
-        if self.nh is None: self.nh = 1552 if self.qrnn else 1152
-        if self.name is None: self.name = self.lang
-
-    @property
-    def tokenizer_prefix(self): return f"{self.tokenizer.value}{self.max_vocab // 1000}k{self.tokenizer_mod}"
-
-    @property
-    def model_direction(self):
-        if self.bidir:
-            return 'bi'
-        if self.backwards:
-            return 'bwd'
-        else:
-            return ''
-
-    @property
-    def model_prefix(self): return self.model_direction + ('qrnn' if self.qrnn else 'lstm')
-
-    @property
-    def model_name(self): return f"{self.model_prefix}_{self.name}{self.model_suffix}.m"
-
-    @property
-    def model_suffix(self): return '' if self.lmseed is None else f'_lmseed-{self.lmseed}'
-
-    @property
-    def pretrained_fnames(self): return [self.base_lm_path / LM_BEST, self.base_lm_path / '../itos'] if self.base_lm_path else None
-
-    def tokenizer_to_fastai_args(self, sp_data_func, use_moses):
-        moses_preproc = [MosesPreprocessingFunc(self.lang)] if use_moses else []
-        if self.tokenizer is Tokenizers.FASTAI_SUBWORD:
-            if self.base_lm_path and not (self.cache_dir / "spm.model").exists():  # ensure we are using the same sentence piece model
-                shutil.copy(self.base_lm_path / '..' / 'spm.model', self.cache_dir)
-                shutil.copy(self.base_lm_path / '..' / 'spm.vocab', self.cache_dir)
-            args = get_sentencepiece_fastai(
-                    cache_dir=self.cache_dir,
-                    vocab_size=self.max_vocab,
-                    lang=self.lang,
-                    pre_rules=moses_preproc + defaults.text_pre_rules)
-
-        elif self.tokenizer is Tokenizers.SUBWORD:
-            if self.base_lm_path and not(self.cache_dir/"spm.model").exists(): # ensure we are using the same sentence piece model
-                shutil.copy(self.base_lm_path / '..' / 'itos.pkl', self.cache_dir)
-                shutil.copy(self.base_lm_path / '..' / 'spm.model', self.cache_dir)
-                shutil.copy(self.base_lm_path / '..' / 'spm.vocab', self.cache_dir)
-            args = get_sentencepiece(self.cache_dir,
-                                     sp_data_func,
-                                     vocab_size=self.max_vocab,
-                                     lang=self.lang,
-                                     pre_rules=moses_preproc + defaults.text_pre_rules,
-                                     post_rules=defaults.text_post_rules,
-                                     fixed_character_coverage=(self.tokenizer_mod=='-fix'))
-        elif self.tokenizer is Tokenizers.MOSES:
-            args = dict(tokenizer=Tokenizer(tok_func=BaseTokenizer,
-                                            lang=self.lang,
-                                            pre_rules=moses_preproc,
-                                            post_rules=[]))
-        elif self.tokenizer is Tokenizers.MOSES_FA:
-            args = dict(tokenizer=Tokenizer(tok_func=BaseTokenizer,
-                                            lang=self.lang,
-                                            pre_rules=moses_preproc + defaults.text_pre_rules,
-                                            post_rules=defaults.text_post_rules))
-        elif self.tokenizer is Tokenizers.FASTAI:
-            args = dict()
-        else:
-            raise ValueError(
-                f"self.tokenizer has wrong value {self.tokenizer}, Allowed values are taken from {Tokenizers}")
-        return args
-
-    def set_seed(self, seed, name):
-        if seed is not None:
-            print(f"Setting {name} seed to {seed}")
-            torch.manual_seed(seed)
-            torch.backends.cudnn.deterministic = True
-            torch.backends.cudnn.benchmark = False
-            np.random.seed(seed)
-
-    def save_info(self):
-        from dataclasses import asdict
-        vals = {k: (str(v) if isinstance(v, Path) else v) for k,v in asdict(self).items()}
-        vals.pop('name', None)
-        vals.pop('lang', None)
-        vals['tokenizer'] = self.tokenizer.value
-        with (self.model_path / 'info.json').open("w") as fp: json.dump(vals, fp)
-        print("Saving info", self.model_path / 'info.json')
-
-    def train_lm(self, num_epochs=20, data_lm=None, bs=70, true_wd=False, drop_mult=0.0, label_smoothing_eps=0.0):
-        print("Training lm")
-        print('Max vocab:', self.max_vocab)
-        print('Cache dir:', self.cache_dir)
-        print('Model dir:', self.model_path)
-        if self.pretrained_fnames or self.pretrained_model:
-            self.set_seed(self.ftseed, "fine-tune")
-        else:
-            self.set_seed(self.lmseed, "LM")
-
-        self.model_path.mkdir(exist_ok=True, parents=True)
-        data_lm = self.load_wiki_data(bs=bs) if data_lm is None else data_lm
-        learn = self.create_lm_learner(data_lm, drop_mult=drop_mult, label_smoothing_eps=label_smoothing_eps)
-        print("Bptt", data_lm.bptt)
-        learn.true_wd = true_wd
-        #learn = learn.to_fp16()
-        if num_epochs > 0:
-            if self.pretrained_fnames or self.pretrained_model:
-                print("Training lm from: ", self.pretrained_fnames or self.pretrained_model)
-                if learn.true_wd:
-                    learn.freeze_to(-1)
-                    learn.fit_one_cycle(1, 1e-2, moms=(0.8, 0.7))
-                    learn.unfreeze()
-                    learn.fit_one_cycle(num_epochs, 1e-3, moms=(0.8, 0.7))
-                else:
-                    learn.freeze_to(-1)
-                    learn.fit_one_cycle(1, 1e-2, moms=(0.8, 0.7), wd=1e-7)  # TODO Fix the learning rates
-                    learn.unfreeze()
-                    learn.fit_one_cycle(num_epochs, 1e-3, moms=(0.8, 0.7), wd=1e-7)
-            else:
-                print("Training lm from random weights")
-                learn.unfreeze()
-                if not learn.true_wd: learn.fit_one_cycle(num_epochs, 5e-3, (0.8, 0.7), wd=1e-7)
-                else:                 learn.fit_one_cycle(num_epochs, 5e-3, (0.8, 0.7)) # TODO find proper values
-        learn.save("lm_best_with_opt", with_opt=True)
-        learn.save_encoder(ENC_BEST)
-        learn.save(LM_BEST, with_opt=False)
-        print(learn.path)
-
-        self.save_info()
-        # do we need to return `learn'? it adds noise to Fire output
-        #return learn
-
-    def create_lm_learner(self, data_lm, dps=None, label_smoothing_eps=0.0, **kwargs):
-        assert self.bidir == False, "bidirectional model is not yet supported"
-        config = awd_lstm_lm_config.copy()
-        config.update(emb_sz=self.emb_sz, n_hid=self.nh, n_layers=self.nl, qrnn=self.qrnn,
-                      tie_weights=True, out_bias=self.out_bias)
-        if dps is not None:
-            config.update(dps)
-
-        trn_args = dict(clip=self.clip, alpha=self.rnn_alpha, beta=self.rnn_beta)
-        trn_args.update(kwargs)
-        print ("Training args: ", trn_args, "config: ", config)
-        learn = language_model_learner(data_lm, AWD_LSTM, config=config, model_dir=self.model_path.relative_to(data_lm.path), pretrained=False, **trn_args)
-        if self.pretrained_model is not None:
-            print("Loading pretrained model")
-            model_path = untar_data(self.pretrained_model, data=False)
-            fnames = [list(model_path.glob(f'*.{ext}'))[0] for ext in ['pth', 'pkl']]
-            learn.load_pretrained(*fnames)
-            learn.freeze()
-        if self.pretrained_fnames is not None:
-            print("Loading pretrained model")
-            fnames = [f'{fn}.{ext}' for fn,ext in zip(self.pretrained_fnames, ['pth', 'pkl'])]
-            learn.load_pretrained(*fnames)
-            learn.freeze()
-        # compared to standard Adam, we set beta_1 to 0.8
-        learn.opt_fn = partial(optim.Adam, betas=(0.8, 0.99))
-        learn.callback_fns += [partial(CSVLogger, filename=f"{learn.model_dir}/lm-history")]
-        if label_smoothing_eps > 0.0:
-            learn.loss_func = FlattenedLoss(LabelSmoothingCrossEntropy, eps=label_smoothing_eps)
-        return learn
-
-    def load_train_text(self):
-        trn_path = self.dataset_path / f'{self.lang}.wiki.train.tokens'
-        with open(trn_path) as f:
-            return [line.rstrip('\n') for line in f]
-
-    def load_wiki_data(self, bs=70):
-        self.model_path.mkdir(exist_ok=True, parents=True)
-        trn_path = self.dataset_path / f'{self.lang}.wiki.train.tokens'
-        val_path = self.dataset_path / f'{self.lang}.wiki.valid.tokens'
-        tst_path = self.dataset_path / f'{self.lang}.wiki.test.tokens'
-        for path_ in [trn_path, val_path, tst_path]:
-            assert path_.exists(), f'Error: {path_} does not exist.'
-
-        args = self.tokenizer_to_fastai_args(sp_data_func=self.load_train_text, use_moses=False)
-
-        data_lm = self.lm_databunch("lm",
-                          train_df=read_wiki_articles(trn_path),
-                          valid_df=read_wiki_articles(val_path),
-                          classes=None,
-                          bs=bs,
-                          text_cols='texts',
-                          bptt=self.bptt,
-                          **args)
-
-        itos, stoi, trn_path = data_lm.vocab.itos, data_lm.vocab.stoi, data_lm.path
-        print('Size of vocabulary:', len(itos))
-        print('First 20 words in vocab:', data_lm.vocab.itos[:20])
-        return data_lm
-
-    def lm_databunch(self, name, *args, **kwargs):
-        return self.databunch(name, bunch_class=TextLMDataBunch, *args, **kwargs)
-
-    def databunch(self, name, bunch_class, train_df, valid_df, bs, force=False, **args):
-        bunch_path = self.cache_dir / name
-        if force and bunch_path.exist():
-            print("Forcefully recreating the databunch, removing previously stored data")
-            for f in bunch_path.glob("*.npy"):
-                f.unlink()
-            if bunch_path.isdir():
-                if name != ".":
-                    bunch_path.rmdir()
-            else:
-                bunch_path.unlink()
-
-        if (bunch_path / 'itos.pkl').exists():
-            data = bunch_class.load(self.cache_dir, name, bs=bs)
-        elif bunch_path.exists():
-            data = load_data(self.cache_dir, name, bs=bs)
-        else:
-            print(f"Running tokenization {name}...")
-            data = make_data_bunch_from_df(cls=bunch_class, path=self.cache_dir,
-                                           train_df=train_df,
-                                           valid_df=valid_df,
-                                           max_vocab=self.max_vocab,
-                                           bs=bs,
-                                           **args)
-
-            data.save(name)
-        with open(self.cache_dir/"itos.pkl", 'wb') as f:
-            pickle.dump(data.vocab.itos, f)
-
-
-        print(f"Data {name}, trn: {len(data.train_ds)}, val: {len(data.valid_ds)}")
-        return data
-
-    @classmethod
-    def from_lm(cls, dataset_path, base_lm_path, **kwargs) -> 'LMHyperParams':
-        dataset_path = Path(dataset_path).resolve()
-        base_lm_path = Path(base_lm_path).resolve()
-        with open(base_lm_path/'info.json', 'r') as f: d = json.load(f)
-        d['dataset_path'] = dataset_path
-        d['base_lm_path'] = base_lm_path
-        d.pop('bs', None)
-        d.pop('drop_mult', None)
-        subword = d.pop('subword', False)
-        tokenizer = d.pop('tokenizer', None)
-        if tokenizer is not None:
-            d['tokenizer'] = Tokenizers(tokenizer)
-        elif subword:
-            d['tokenizer'] = Tokenizers.SUBWORD
-        else:
-            d['tokenizer'] = Tokenizers.MOSES
-
-        d.update(kwargs)
-        return cls(**d)
-
-    @classmethod
-    def from_json(cls, model_path:Path, **kwargs):
-        model_path = Path(model_path).resolve()
-        name = folder_name_to_model_name(model_path)
-        with open(model_path / 'info.json', 'r') as f:
-            d = json.load(f)
-        d.update(kwargs)
-        d['name'] = name
-        dataset_path = path_strip(model_path, "data", "models").parent
-        d['dataset_path'] = str(dataset_path)
-        d['lang'] = infer_lang_from_dataset(dataset_path.name)
-        return cls(**d)
-
-
-    def resolve_template(self, template, **additional_options):
-        return super().resolve_template(template, model_dir=self.model_path, **additional_options)
-
-def infer_lang_from_dataset(name:str):
-    return name.split("-")[0]
-
-def path_strip(path, from_folder, to_folder):
-    to_p = [p for p in path.parents if p.name == to_folder][0]
-    from_p = [p for p in path.parents if p.name == from_folder][0]
-    return to_p.relative_to(from_p.parent)
-
-def validate_lm(self):
-    if not self.exp.subword and self.exp.max_vocab is None:
-        raise NotImplementedError("figure out how to validate and save results")
-        # only if we use the unpreprocessed version and the full vocabulary
-        # are the perplexity results comparable to previous work
-        print(f"Validating model performance with test tokens from: {trn_path}")
-        tst_tok = read_whitespace_file(trn_path)
-        tst_ids = np.array([([stoi.get(w, stoi[UNK]) for w in s]) for s in tst_tok])
-        logloss, perplexity = validate(learn.model, tst_ids, self.exp.bptt)
-        print('Test logloss:', logloss.item(), 'perplexity:', perplexity.item())
-
-def get_data_folder() -> Path:
-    """
-    return data folder to use for future processing
-    """
-    return (pathlib.Path(__file__).parent.parent / "data")
